+++ packages/fast-check/test/unit/arbitrary/__test-helpers__/ArbitraryHelpers.ts
@@ -17,7 +17,6 @@ export function fakeArbitraryClass<T = any>(): { Class: new () => Arbitrary<T> }
   const map = vi.fn();
   const chain = vi.fn();
   const noShrink = vi.fn();
-  const noBias = vi.fn();
 
   class FakeArbitrary extends Arbitrary<T> {
     generate = generate;
@@ -27,9 +26,8 @@ export function fakeArbitraryClass<T = any>(): { Class: new () => Arbitrary<T> }
     map = map;
     chain = chain;
     noShrink = noShrink;
-    noBias = noBias;
   }
-  return { Class: FakeArbitrary, generate, canShrinkWithoutContext, shrink, filter, map, chain, noShrink, noBias };
+  return { Class: FakeArbitrary, generate, canShrinkWithoutContext, shrink, filter, map, chain, noShrink };
 }
 
 /**
