+++ packages/fast-check/test/unit/check/arbitrary/definition/Arbitrary.utest.spec.ts
@@ -734,52 +734,4 @@ describe('NextArbitrary', () => {
       expect(secondNoShrink).toBe(firstNoShrink);
     });
   });
-
-  describe('noBias', () => {
-    it('should override passed bias with undefined', () => {
-      // Arrange
-      const generate = vi.fn();
-      class MyNextArbitrary extends Arbitrary<any> {
-        generate = generate;
-        canShrinkWithoutContext(value: unknown): value is any {
-          throw new Error('Not implemented.');
-        }
-        shrink(): Stream<Value<any>> {
-          throw new Error('Not implemented.');
-        }
-      }
-      const fakeArbitrary: Arbitrary<any> = new MyNextArbitrary();
-      const noBiasArbitrary = fakeArbitrary.noBias();
-
-      // Act
-      noBiasArbitrary.generate(mrngNoCall, 42);
-
-      // Assert
-      expect(generate).toHaveBeenCalledTimes(1);
-      expect(generate).toHaveBeenCalledWith(mrngNoCall, undefined);
-    });
-
-    it('should return itself when called twice', () => {
-      // Arrange
-      class MyNextArbitrary extends Arbitrary<any> {
-        generate(): Value<any> {
-          throw new Error('Not implemented.');
-        }
-        canShrinkWithoutContext(value: unknown): value is any {
-          throw new Error('Not implemented.');
-        }
-        shrink(): Stream<Value<any>> {
-          throw new Error('Not implemented.');
-        }
-      }
-      const fakeArbitrary: Arbitrary<any> = new MyNextArbitrary();
-
-      // Act
-      const firstNoBias = fakeArbitrary.noBias();
-      const secondNoBias = firstNoBias.noBias();
-
-      // Assert
-      expect(secondNoBias).toBe(firstNoBias);
-    });
-  });
 });
