+++ website/docs/core-blocks/arbitraries/combiners/any.md
@@ -306,26 +306,6 @@ fc.nat().chain((min) => fc.tuple(fc.constant(min), fc.integer({ min, max: 0xffff
 Resources: [API reference](https://fast-check.dev/api-reference/classes/Arbitrary.html#chain).  
 Available since 1.2.0.
 
-## .noBias
-
-Drop bias from an existing arbitrary. Instead of being more likely to generate certain values the resulting arbitrary will be close to an equi-probable generator.
-
-**Signatures:**
-
-- `.noBias()` — _deprecated since v3.20.0 ([#5050](https://github.com/dubzzz/fast-check/pull/5050))_
-
-**Usages:**
-
-```js
-fc.nat().noBias();
-// Note: Compared to fc.nat() alone, the generated values are evenly distributed in
-// the range 0 to 0x7fffffff making small values much more unlikely.
-// Examples of generated values: 422394692, 1060515252, 383444404, 1509445429, 659009523…
-```
-
-Resources: [API reference](https://fast-check.dev/api-reference/classes/Arbitrary.html#noBias).  
-Available since 1.1.0.
-
 ## .noShrink
 
 Drop shrinking capabilities from an existing arbitrary.
